lab02/post_service/requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
python-multipart
pydantic>=2.0
requests